fix: detecting processor

+++ scripts/afterPack.js
@@ -1,32 +1,40 @@
 const fs = require("fs-extra");
 const path = require("path");
+const { Arch } = require("builder-util");
 
 exports.default = async function (context) {
   if (context.electronPlatformName !== "darwin") return;
 
-  const appOutDir = context.appOutDir;
-  const projectDir = context.projectDir || path.resolve(__dirname, "..");
-
-  let archFolder;
-  if (context.arch === "arm64") {
-    archFolder = "aarch64";
-  } else if (context.arch === "x64") {
-    archFolder = "x86_64";
-  } else {
-    console.log("Do nothing for architecture:", context.arch);
+  if (context.arch === Arch.universal) {
+    console.log("Skipping Tor copy during universal merge stage");
     return;
   }
 
-  const resourcesPath = path.join(appOutDir, "Bastyon.app", "Contents", "Resources");
-  const torDest = path.join(resourcesPath, "tor", archFolder);
-  const torSrc = path.join(projectDir, "tor", "macos", archFolder);
+  let archFolder;
+  if (context.arch === Arch.arm64) archFolder = "aarch64";
+  else if (context.arch === Arch.x64) archFolder = "x86_64";
+  else {
+    console.log("Unsupported arch:", context.arch);
+    return;
+  }
 
-  console.log("Removing old Tor folder (if exists):", torDest);
-  await fs.remove(torDest);
+  const appOutDir = context.appOutDir;
+  const projectDir = context.projectDir || path.resolve(__dirname, "..");
+
+  const resourcesPath = path.join(appOutDir, "Bastyon.app", "Contents", "Resources");
+  const torSrc = path.join(projectDir, "tor", "macos", archFolder);
+  const torDest = path.join(resourcesPath, "tor", archFolder);
+
+  if (!await fs.pathExists(torSrc)) {
+    console.warn("Tor source missing:", torSrc);
+    return;
+  }
 
   console.log(`Copying ${archFolder} Tor ->`, torDest);
+
+  await fs.remove(torDest);
   await fs.copy(torSrc, torDest);
 
-  console.log("Tor copied successfully");
+  console.log(`${archFolder} Tor copied`);
 };
